Make linter happier
